Move OpUtils to SciJava Ops Engine
The things that we needed in SciJava Ops API were better suited to live
in the classes that used them. With those things moved, there was no
reason to keep OpUtils in SciJava Ops API.
