+++ scijava/scijava-ops-engine/src/main/java/org/scijava/ops/engine/struct/ClassParameterMemberParser.java
@@ -9,7 +9,7 @@ import java.util.Arrays;
 import java.util.List;
 
 import org.scijava.ValidityProblem;
-import org.scijava.ops.api.OpUtils;
+import org.scijava.ops.engine.OpUtils;
 import org.scijava.struct.MemberParser;
 import org.scijava.struct.Structs;
 import org.scijava.struct.ValidityException;
